Tags directory within DataModel.
Also namespace changes
