+++ Minio.Examples/Cases/PutObjectWithTags.cs
@@ -14,10 +14,11 @@
  * limitations under the License.
  */
 
-using Minio.DataModel;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+
+using Minio.DataModel.Tags;
 
 namespace Minio.Examples.Cases
 {
