+++ Minio.Examples/Cases/SetObjectTags.cs
@@ -1,5 +1,5 @@
 /*
- * MinIO .NET Library for Amazon S3 Compatible Cloud Storage, (C) 2020 MinIO, Inc.
+ * MinIO .NET Library for Amazon S3 Compatible Cloud Storage, (C) 2020, 2021 MinIO, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,7 +16,7 @@
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using Minio.DataModel;
+using Minio.DataModel.Tags;
 
 namespace Minio.Examples.Cases
 {
