+++ Minio/ApiEndpoints/ObjectOperations.cs
@@ -21,7 +21,6 @@ using System.Collections.Generic;
 using System.Globalization;
 using System.IO;
 using System.Linq;
-using System.Net;
 using System.Reactive.Linq;
 using System.Threading;
 using System.Threading.Tasks;
@@ -29,6 +28,7 @@ using System.Xml.Linq;
 using System.Xml.Serialization;
 
 using Minio.DataModel;
+using Minio.DataModel.Tags;
 using Minio.Exceptions;
 using Minio.Helper;
