+++ Minio/DataModel/BucketOperationsArgs.cs
@@ -20,6 +20,7 @@ using System.IO;
 using Minio.DataModel;
 using Minio.DataModel.ILM;
 using Minio.DataModel.Replication;
+using Minio.DataModel.Tags;
 using Minio.Exceptions;
 using RestSharp;
