+++ Minio/DataModel/ILM/RuleFilter.cs
@@ -16,6 +16,7 @@
 
 using System;
 using System.Xml.Serialization;
+using Minio.DataModel.Tags;
 
 /*
  * RuleFilter class used within LifecycleRule which encapsulates filter information.
