+++ Minio/DataModel/ObjectOperationsResponse.cs
@@ -23,6 +23,7 @@ using System.Net;
 using System.Xml.Serialization;
 
 using Minio.DataModel;
+using Minio.DataModel.Tags;
 using RestSharp;
 
 namespace Minio
